Generalizing the NextHeaderDescription to be a BaseHeader for all headers
Following the discussions in the CWG4 meeting and in an attempt to merge
the proposal by the DAQ group for a generalized header structure, the
NextHeaderDescription has now been transferred into a basic header structure
with the minimum amount of common information.

The DataHeader structure inherits from the new BaseHeader structure.